Added various curve25519 testvectors (rfc7748/gcrypt)
Found several of the test vectors failed because we didn't clear the MSB
bit in the ucoord. According to RFC7748 section 5,

    When receiving such an array, implementations of X25519 (but not
    X448) MUST mask the most significant bit in the final byte.  This is
    done to preserve compatibility with point formats that reserve the
    sign bit for use in other protocols and to increase resistance to
    implementation fingerprinting.

Changing this doesn't affect our main use of this function, generating
public/private keypairs for WireGuard because there the base coordinate
is 9 and doesn't have the MSB bit set anyway.